requirements.txt
mysql-connector-python
matplotlib
python-dotenv